test: move JenkinsPayload string to own class
