+++ src/test/java/org/terasology/launcher/repositories/JenkinsRepositoryAdapterTest.java
@@ -41,89 +41,12 @@ class JenkinsRepositoryAdapterTest {
     @BeforeAll
     static void setup() throws MalformedURLException {
         gson = new Gson();
-        validResult = gson.fromJson(validPayload(), Jenkins.ApiResult.class);
-        incompleteResults = incompletePayloads().stream()
+        validResult = gson.fromJson(JenkinsPayload.validPayload(), Jenkins.ApiResult.class);
+        incompleteResults = JenkinsPayload.incompletePayloads().stream()
                 .map(json -> gson.fromJson(json, Jenkins.ApiResult.class))
                 .collect(Collectors.toList());
         expectedArtifactUrl = new URL("http://jenkins.terasology.io/teraorg/job/Nanoware/job/Omega/job/develop/1/"
                 + "artifact/" + "distros/omega/build/distributions/TerasologyOmega.zip");
-    }
-
-    static String validPayload() {
-        return "{\n" +
-                "  \"builds\": [\n" +
-                "    {\n" +
-                "      \"artifacts\": [\n" +
-                "        {\n" +
-                "          \"fileName\": \"TerasologyOmega.zip\",\n" +
-                "          \"relativePath\": \"distros/omega/build/distributions/TerasologyOmega.zip\"\n" +
-                "        },\n" +
-                "        {\n" +
-                "          \"fileName\": \"versionInfo.properties\",\n" +
-                "          \"relativePath\": \"distros/omega/versionInfo.properties\"\n" +
-                "        }\n" +
-                "      ],\n" +
-                "      \"number\": 1,\n" +
-                "      \"result\": \"SUCCESS\",\n" +
-                "      \"timestamp\": 1604285977306,\n" +
-                "      \"url\": \"http://jenkins.terasology.io/teraorg/job/Nanoware/job/Omega/job/develop/1/\"\n" +
-                "    }\n" +
-                "  ]\n" +
-                "}";
-    }
-
-    static String nullArtifactsPayload() {
-        return "{ \n" +
-                "  \"builds\": [\n" +
-                "    {\n" +
-                "      \"number\": 1, \"result\": \"SUCCESS\", \"timestamp\": 1604285977306, \n" +
-                "      \"url\": \"http://jenkins.terasology.io/teraorg/job/Nanoware/job/Omega/job/develop/1/\"\n" +
-                "    }\n" +
-                "  ]\n" +
-                "}";
-    }
-
-    static String emptyArtifactsPayload() {
-        return "{\n" +
-                "  \"builds\": [\n" +
-                "    {\n" +
-                "      \"artifacts\": [],\n" +
-                "      \"number\": 1,\n" +
-                "      \"result\": \"SUCCESS\",\n" +
-                "      \"timestamp\": 1604285977306,\n" +
-                "      \"url\": \"http://jenkins.terasology.io/teraorg/job/Nanoware/job/Omega/job/develop/1/\"\n" +
-                "    }\n" +
-                "  ]\n" +
-                "}";
-    }
-
-    static String incompleteArtifactsPayload() {
-        return "{\n" +
-                "  \"builds\": [\n" +
-                "    {\n" +
-                "      \"artifacts\": [\n" +
-                "        {\n" +
-                "          \"fileName\": \"versionInfo.properties\",\n" +
-                "          \"relativePath\": \"distros/omega/versionInfo.properties\"\n" +
-                "        }\n" +
-                "      ],\n" +
-                "      \"number\": 1,\n" +
-                "      \"result\": \"SUCCESS\",\n" +
-                "      \"timestamp\": 1604285977306,\n" +
-                "      \"url\": \"http://jenkins.terasology.io/teraorg/job/Nanoware/job/Omega/job/develop/1/\"\n" +
-                "    }\n" +
-                "  ]\n" +
-                "}";
-    }
-
-    static List<String> incompletePayloads() {
-        return List.of(
-                "{}",
-                "{ \"builds\": [] }",
-                nullArtifactsPayload(),
-                emptyArtifactsPayload(),
-                incompleteArtifactsPayload()
-        );
     }
 
     static Stream<Arguments> incompleteResults() {
